Updated todo list app with new features

+++ controllers/task.js
@@ -1,26 +1,32 @@
+import ErrorHandler from "../middleware/error.js";
 import { Task } from "../models/task.js";
 
 
 export const newTask=async (req,res,next)=>{
+    try {
+        const {title,description}=req.body;
 
-    const {title,description}=req.body;
-
-    await Task.create({
-        title,
-        description,
-        user:req.user,
-    })
-
-    res.status(201).json({
-        success:true,
-        message:"Task added successfully"
-    })
+        await Task.create({
+            title,
+            description,
+            user:req.user,
+        })
+    
+        res.status(201).json({
+            success:true,
+            message:"Task added successfully"
+        })
+    
+    } catch (error) {
+        next(error)
+    }
 
 }
 
 
 export const getMyTask = async (req,res,next)=>{
 
+try {
     const userid =req.user._id;
 
     const tasks= await Task.find({user: userid});
@@ -29,18 +35,19 @@ export const getMyTask = async (req,res,next)=>{
         success:true,
         tasks
     })
+} catch (error) {
+    next(error)
+}
 
 }
 
 export const updateTask = async (req,res,next)=>{
+try {
     const {id}=req.params;
     const task= await Task.findById(id);
 
-    if(!task)
-        return res.status(404).json({
-            success:true,
-            message:"Invalid Id"
-        })
+    if(!task) return next(new ErrorHandler("task not found",404))
+
 
     task.isCompleted=!task.isCompleted;
 
@@ -52,11 +59,17 @@ export const updateTask = async (req,res,next)=>{
         message:"Task updated"
 
     })
+} catch (error) {
+    next(error)
+}
 
 }
 
 export const deleteTask = async (req,res,next)=>{
+try {
     const task = await Task.findById(req.params.id);
+
+    if(!task) return next(new ErrorHandler("task not found",404))
 
     await task.deleteOne();
 
@@ -66,5 +79,8 @@ export const deleteTask = async (req,res,next)=>{
         message:"Task deleted"
 
     })
+} catch (error) {
+    next(error)
+}
 
 }
